+++ src/services/MessageModalService.js
@@ -1,5 +1,5 @@
-import { store } from 'StateLayer'
-import { open, close } from 'redux/messageModal'
+import store from 'redux/store'
+import { open, close, reset } from 'redux/messageModal'
 
 import { MessageType } from 'components/modal/message'
 
@@ -7,10 +7,16 @@ export class MessageModalService {
 	dispatch = null
 
 	constructor() {
+		this.setDispatch()
+	}
+
+	setDispatch() {
 		this.dispatch = store.dispatch
 	}
 
 	async open(text, type) {
+		if (!this.dispatch) this.setDispatch()
+
 		this.dispatch(open({ text, type }))
 	}
 
@@ -32,5 +38,9 @@ export class MessageModalService {
 
 	async close() {
 		this.dispatch(close())
+
+		setTimeout(() => {
+			this.dispatch(reset())
+		}, 1000)
 	}
 }