+++ src/services/base/index.js
@@ -1,1 +1,2 @@
 export * from './ServiceBase'
+export * from './ReduxService'